requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
mesa>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
